webService/requirements.txt
alembic==1.13.1
attrs==23.2.0
bidict==0.23.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-Reuploaded==1.2.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.3
gevent==22.10.2
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.7
importlib_metadata==7.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysqlclient==2.1.1
packaging==24.0
pluggy==1.4.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.8.0
pytest==7.0.1
python-dotenv==0.21.1
python-engineio==4.9.0
python-socketio==5.11.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.28
tomli==2.0.1
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==2.2.3
wsproto==1.2.0
zipp==3.18.1
zope.event==5.0
zope.interface==6.2